requirements.txt
flask>=2.0.0
opencv-python
pytesseract
sympy
numpy
Pillow